Moved ButtonIcon inside a folder ButtonIcon
This seems to be the common practice when we want to add a new variation (e.g. the ButtonIconDanger)
