+++ lib/ButtonNew/ButtonIcon/ButtonIcon.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { omit } from 'lodash';
 import { string, bool } from 'prop-types';
 import cx from 'classnames';
-import { ButtonBase } from './ButtonBase';
-import Icon from '../Icon';
-import { propTypes, defaultProps, sizes } from './common';
+import { ButtonBase } from '../ButtonBase';
+import Icon from '../../Icon';
+import { propTypes, defaultProps, sizes } from '../common';
 
 function ButtonIcon({ className, disabled, name, active, size, ...rest }) {
   const classes = cx(
