+++ lib/ButtonNew/stories/ButtonIcon/ButtonIcon.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { ButtonIcon } from '../ButtonIcon';
-import StoryItem from '../../../stories/styleguide/StoryItem';
-import { ButtonStoryItem } from './ButtonStoryItem';
+import { ButtonIcon } from '../../ButtonIcon/ButtonIcon';
+import StoryItem from '../../../../stories/styleguide/StoryItem';
+import { ButtonStoryItem } from '../ButtonStoryItem';
 
 storiesOf('Components', module).add('ButtonIcon', () => (
   <StoryItem title="ButtonIcon" description="The icon button component">
