+++ lib/Toolbar/stories/Toolbar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import StoryItem from '../../../stories/styleguide/StoryItem';
 import { Toolbar } from '../Toolbar';
-import { ButtonIcon } from '../../ButtonNew/ButtonIcon';
+import { ButtonIcon } from '../../ButtonNew/ButtonIcon/ButtonIcon';
 
 storiesOf('Components', module).add('Toolbar', () => (
   <StoryItem title="Toolbar" description="The toolbar component">
