+++ lib/index.js
@@ -130,6 +130,6 @@ export { ButtonTertiary } from './ButtonNew/ButtonTertiary/ButtonTertiary';
 export {
   ButtonTertiaryDanger
 } from './ButtonNew/ButtonTertiary/ButtonTertiaryDanger';
-export { ButtonIcon } from './ButtonNew/ButtonIcon';
+export { ButtonIcon } from './ButtonNew/ButtonIcon/ButtonIcon';
 export { Toolbar } from './Toolbar/Toolbar';
 export { OptionMenu } from './OptionMenu/OptionMenu';
