+++ stories/index.js
@@ -72,6 +72,6 @@ import '../lib/ButtonNew/stories/ButtonPrimary/ButtonPrimary'
 import '../lib/ButtonNew/stories/ButtonPrimary/ButtonPrimaryDanger'
 import '../lib/ButtonNew/stories/ButtonTertiary/ButtonTertiary'
 import '../lib/ButtonNew/stories/ButtonTertiary/ButtonTertiaryDanger'
-import '../lib/ButtonNew/stories/ButtonIcon'
+import '../lib/ButtonNew/stories/ButtonIcon/ButtonIcon'
 import '../lib/Toolbar/stories/Toolbar'
 import '../lib/OptionMenu/stories/OptionMenu'
